PMM-6372 Refactoring
