+++ pmm-app/src/pmm-settings/components/SignUp/FieldAdapters/Checkbox.tsx
