+++ pmm-app/src/pmm-settings/components/SignUp/LoggedIn/LoggedIn.constants.ts
@@ -1,0 +1,1 @@
+export const SIGN_OUT_URL = '#';
